Compile fix

+++ src/strife/info.cpp
@@ -3520,9 +3520,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] =
         1000,       //mass
         0,      //damage
         sfx_alnact,     //activesound
-        MF_SPECIAL|MF_SOLID|MF_SHOOTABLE|MF_NOGRAVITY|MF_GIVEQUEST
-        |MF_FLOAT|MF_NODIALOG|MF_SHADOW|MF_COUNTKILL|MF_NOTDMATCH
-        |MF_MVIS|MF_SPECTRAL,        //flags
+        static_cast<unsigned int>(MF_SPECIAL|MF_SOLID|MF_SHOOTABLE|MF_NOGRAVITY|MF_GIVEQUEST|MF_FLOAT|MF_NODIALOG|MF_SHADOW|MF_COUNTKILL|MF_NOTDMATCH|MF_MVIS|MF_SPECTRAL),        //flags
         nullptr,       //namepointer
     },
 
@@ -3856,7 +3854,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] =
         100,        //mass
         10,     //damage
         sfx_None,       //activesound
-        MF_NOBLOCKMAP|MF_NOGRAVITY|MF_DROPOFF|MF_MISSILE|MF_SPECTRAL,       //flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP|MF_NOGRAVITY|MF_DROPOFF|MF_MISSILE|MF_SPECTRAL),       //flags
         nullptr,       //namepointer
     },
 
@@ -3910,7 +3908,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] =
         100,        //mass
         130,        //damage
         sfx_None,       //activesound
-        MF_NOBLOCKMAP|MF_NOGRAVITY|MF_DROPOFF|MF_MISSILE|MF_SPECTRAL,       //flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP|MF_NOGRAVITY|MF_DROPOFF|MF_MISSILE|MF_SPECTRAL),       //flags
         nullptr,       //namepointer
     },
 
@@ -3937,7 +3935,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] =
         100,        //mass
         30,     //damage
         sfx_None,       //activesound
-        MF_NOBLOCKMAP|MF_NOGRAVITY|MF_DROPOFF|MF_MISSILE|MF_SPECTRAL,       //flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP|MF_NOGRAVITY|MF_DROPOFF|MF_MISSILE|MF_SPECTRAL),       //flags
         nullptr,       //namepointer
     },
 
@@ -3964,7 +3962,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] =
         100,        //mass
         100,        //damage
         sfx_None,       //activesound
-        MF_NOBLOCKMAP|MF_NOGRAVITY|MF_DROPOFF|MF_MISSILE|MF_SPECTRAL,       //flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP|MF_NOGRAVITY|MF_DROPOFF|MF_MISSILE|MF_SPECTRAL),       //flags
         nullptr,       //namepointer
     },
 
@@ -3991,7 +3989,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] =
         100,        //mass
         50,     //damage
         sfx_None,       //activesound
-        MF_NOBLOCKMAP|MF_NOGRAVITY|MF_DROPOFF|MF_MISSILE|MF_SPECTRAL,       //flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP|MF_NOGRAVITY|MF_DROPOFF|MF_MISSILE|MF_SPECTRAL),       //flags
         nullptr,       //namepointer
     },
 
@@ -4045,7 +4043,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] =
         100,        //mass
         120,        //damage
         sfx_None,       //activesound
-        MF_NOBLOCKMAP|MF_NOGRAVITY|MF_DROPOFF|MF_MISSILE|MF_SPECTRAL,       //flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP|MF_NOGRAVITY|MF_DROPOFF|MF_MISSILE|MF_SPECTRAL),       //flags
         nullptr,       //namepointer
     },
 
@@ -4072,7 +4070,7 @@ mobjinfo_t mobjinfo[NUMMOBJTYPES] =
         100,        //mass
         60,     //damage
         sfx_None,       //activesound
-        MF_NOBLOCKMAP|MF_NOGRAVITY|MF_DROPOFF|MF_MISSILE|MF_SPECTRAL,       //flags
+        static_cast<unsigned int>(MF_NOBLOCKMAP|MF_NOGRAVITY|MF_DROPOFF|MF_MISSILE|MF_SPECTRAL),       //flags
         nullptr,       //namepointer
     },
